pyproject.toml
[tool.poetry]
name = "bananastore"
version = "0.1.0"
description = "Image generation done easy"
license = "MIT"
authors = ["Michael Ikemann"]
readme = "README.md"
repository = "https://github.com/Alyxion/BananaStore"
keywords = ["image-generation", "openai", "google", "gemini", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics",
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.116.1"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
python-dotenv = "^1.1.1"
pillow = "^11.3.0"
python-multipart = "^0.0.20"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
